+++ solr/__init__.py
@@ -13,6 +13,7 @@ from .extract import Extract
 from .response import SpellcheckResult
 from .response import GroupedResult, GroupField, Group
 from .knn import KNN
+from .async_solr import AsyncSolr
 from .field import Field
 from .sort import Sort
 from .facet import Facet